Move grid to whiteboard
